pyproject.toml
[project]
name = "google-drive-mcp"
version = "0.1.0"
description = "MCP server for Google Drive integration with focus on Google Docs operations"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=0.1.0",
    "google-api-python-client>=2.100.0",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.2.0"
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.0.0"
]

[project.scripts]
google-drive-mcp = "google_drive_mcp.server:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=google_drive_mcp",
    "--cov-report=html",
    "--cov-report=term-missing",
    "--strict-markers"
]
asyncio_mode = "auto"
